Add linter to CI pipeline  (#844)

+++ clerk/side_handler.go
@@ -45,10 +45,9 @@ func NewPostTxHandler(k Keeper, contractCaller helper.IContractCaller) hmTypes.P
 }
 
 func SideHandleMsgEventRecord(ctx sdk.Context, k Keeper, msg types.MsgEventRecord, contractCaller helper.IContractCaller) (result abci.ResponseDeliverSideTx) {
-
 	k.Logger(ctx).Debug("✅ Validating External call for clerk msg",
 		"txHash", hmTypes.BytesToHeimdallHash(msg.TxHash.Bytes()),
-		"logIndex", uint64(msg.LogIndex),
+		"logIndex", msg.LogIndex,
 		"blockNumber", msg.BlockNumber,
 	)
 
@@ -86,6 +85,7 @@ func SideHandleMsgEventRecord(ctx sdk.Context, k Keeper, msg types.MsgEventRecor
 			"EventContractAddress", eventLog.ContractAddress.String(),
 			"MsgContractAddress", msg.ContractAddress.String(),
 		)
+
 		return hmCommon.ErrorSideTx(k.Codespace(), common.CodeInvalidMsg)
 	}
 
@@ -97,6 +97,7 @@ func SideHandleMsgEventRecord(ctx sdk.Context, k Keeper, msg types.MsgEventRecor
 					"EventData", hmTypes.BytesToHexBytes(eventLog.Data),
 					"MsgData", hmTypes.BytesToHexBytes(msg.Data),
 				)
+
 				return hmCommon.ErrorSideTx(k.Codespace(), common.CodeInvalidMsg)
 			}
 		} else {
@@ -106,17 +107,18 @@ func SideHandleMsgEventRecord(ctx sdk.Context, k Keeper, msg types.MsgEventRecor
 					"EventData", hmTypes.BytesToHexBytes(eventLog.Data),
 					"MsgData", hmTypes.BytesToHexBytes(msg.Data),
 				)
+
 				return hmCommon.ErrorSideTx(k.Codespace(), common.CodeInvalidMsg)
 			}
 		}
 	}
 
 	result.Result = abci.SideTxResultType_Yes
+
 	return
 }
 
 func PostHandleMsgEventRecord(ctx sdk.Context, k Keeper, msg types.MsgEventRecord, sideTxResult abci.SideTxResultType) sdk.Result {
-
 	// Skip handler if clerk is not approved
 	if sideTxResult != abci.SideTxResultType_Yes {
 		k.Logger(ctx).Debug("Skipping new clerk since side-tx didn't get yes votes")
